\name{idm-package}
\alias{idm-package}
\alias{idm}
\docType{package}
\title{
Incremental Decomposition Methods in R
}
\description{
Incremental Principal Component Analysis and Multiple Correspondence Analysis using incremental eigenvalue decomposition methods
}
\details{
\tabular{ll}{
Package: \tab idm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-10\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Alfonso Iodice D' Enza [aut],
Angelos Markos [aut, cre],
Davide Buttarazzi [ctb]
}
\references{
Hall, P., Marshall, D., & Martin, R. (2002). Adding and subtracting eigenspaces with eigenvalue decomposition and singular value decomposition. \emph{Image and Vision Computing}, 20(13), 1009-1016.\cr
Ross, D. A., Lim, J., Lin, R. S., & Yang, M. H. (2008). Incremental learning for robust visual tracking. \emph{International Journal of Computer Vision}, 77(1-3), 125-141.\cr
Iodice D' Enza, A., & Markos, A. (2014). Low-dimensional tracking of association structures in categorical data. \emph{Statistics and Computing}, (forthcoming).
}

\keyword{ package }
