\name{idm-package}
\alias{idm-package}
\alias{idm}
\docType{package}
\title{
Incremental Decomposition Methods
}
\description{
Incremental Multiple Correspondence Analysis and Principal Component Analysis 
}
\details{
\tabular{ll}{
Package: \tab idm\cr
Type: \tab Package\cr
Version: \tab 1.7.1\cr
Date: \tab 2016-11-01\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Alfonso Iodice D' Enza [aut],
Angelos Markos [aut, cre],
Davide Buttarazzi [ctb]
}
\references{
Hall, P., Marshall, D., & Martin, R. (2002). Adding and subtracting eigenspaces with eigenvalue decomposition and singular value decomposition. \emph{Image and Vision Computing}, \emph{20}(13), 1009-1016.\cr \cr
Ross, D. A., Lim, J., Lin, R. S., & Yang, M. H. (2008). Incremental learning for robust visual tracking. \emph{International Journal of Computer Vision}, \emph{77}(1-3), 125-141.\cr \cr
Iodice D' Enza, A., & Markos, A. (2015). Low-dimensional tracking of association structures in categorical data, \emph{Statistics and Computing}, \emph{25}(5), 1009-1022. 
}

\keyword{ package }
